changes as pull request comments, need to test in Kupe
