+++ im_calc_sl.template
@@ -19,7 +19,7 @@
 date
 
 source machine_env.sh
-export IMPATH=/home/melody.zhu/IM_calculation
+export IMPATH=/nesi/nobackup/nesi00213/IM_calculation
 export PYTHONPATH=/nesi/nobackup/nesi00213/tmp/qcore:/$PYTHONPATH:$IMPATH
 
 echo ___calculating simulations___
